add Boxes component
